Added Custom Exceptions and ControllerAdvice for handle this customs exception and send proper Rest Response

+++ EmployeeManagement/src/main/java/com/evoke/employeemanagement/service/EmployeeServiceImpl.java
@@ -8,17 +8,29 @@ import org.springframework.stereotype.Service;
 
 import com.evoke.employeemanagement.dao.EmployeeRepo;
 import com.evoke.employeemanagement.entity.Employee;
+import com.evoke.employeemanagement.exception.BusinessException;
+import com.evoke.employeemanagement.exception.InvalidEmailException;
+import com.evoke.employeemanagement.exception.ResourceNotFoundException;
 
 @Service
-public class EmployeeServiceImpl implements IEmployeeService{
+public class EmployeeServiceImpl implements IEmployeeService {
 
 	@Autowired
-	private EmployeeRepo employeeRepo; 
-	
+	private EmployeeRepo employeeRepo;
+
 	@Override
 	public Employee addEmployee(Employee employee) {
+		if(!isValid(employee.getEmail())) {
+			throw new InvalidEmailException("Provide proper email");
+		}
 		employeeRepo.save(employee);
 		return employee;
+	}
+	
+	public static boolean isValid(String email) {
+		String regex = "^[a-zA-Z]+@[a-zA-Z.-]+$";
+		boolean val=email.matches(regex);
+		return val;
 	}
 
 	@Override
@@ -28,19 +40,35 @@ public class EmployeeServiceImpl implements IEmployeeService{
 
 	@Override
 	public Employee findEmployeeById(Long id) {
+			Employee emp = employeeRepo.findEmployeeById(id);
+			if (emp == null || emp.getEmpName().isEmpty())
+				throw new ResourceNotFoundException("Resource not found for Id : " + id);
+			return emp;
 		
-		return employeeRepo.findEmployeeById(id);
 	}
 
-	
 	@Override
 	public void deleteEmployee(Long id) {
-		employeeRepo.deleteById(id);
+		
+			Employee emp = employeeRepo.findEmployeeById(id);
+			System.out.println("Hello");
+			if (emp == null) {
+				throw new ResourceNotFoundException("Resource not found for Id : " + id);
+			}
+			employeeRepo.deleteById(id);
 	}
 
 	@Override
 	public Employee updateEmployee(Employee employee) {
-		return employeeRepo.save(employee);
-}
+			Employee empDao = employeeRepo.findEmployeeById(employee.getEmpId());
+			if (empDao == null) {
+				throw new ResourceNotFoundException(
+						"User with employeeId : " + employee.getEmpId() + " Does not exits");
+			}
+			return employeeRepo.save(employee);
+
+		}
+		
+	
 
 }
